Add files via upload
Modified add to map function to be simpler and reduced redundancies. Also added feature to create multiple orders by passing in the int param num_orders in the function call. This allows for autonomous creation of orders/records. Also Added list comprehension in the create_barcode_hashes function to create a list of tuples